Refactoring the web driver functional tests to use the page model as they were originally intended to.
