+++ tests/ui/landing_page_test.py
@@ -2,8 +2,6 @@
 import unittest
 
 import flask
-from selenium.webdriver.common.action_chains import ActionChains
-from selenium.webdriver.common.by import By
 from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.support.ui import WebDriverWait
 
@@ -25,8 +23,13 @@ class LandingPageTest(BaseTest):
 
   def tearDown(self):
     """Teardown for test methods."""
-    self.removeTestUser(should_raise_exception=False)
-    self.removeTestServer(should_raise_exception=False)
+    landing_page = LandingPage(self.driver)
+    landing_page.removeTestUser(BaseTest.TEST_USER_AS_DICT['name'],
+                                self.args.server_url,
+                                should_raise_exception=False)
+    landing_page.removeTestServer(BaseTest.TEST_SERVER_AS_DICT['name'],
+                                  self.args.server_url,
+                                  should_raise_exception=False)
     LoginPage(self.driver).Logout(self.args.server_url)
     super(LandingPageTest, self).tearDown()
 
@@ -53,45 +56,62 @@ class LandingPageTest(BaseTest):
 
   def testManuallyAddUserFromLandingPage(self):
     """Test that manually adding a user shows up on the user listing."""
+    landing_page = LandingPage(self.driver)
     self.assertTestUserPresenceOnPage(False)
 
-    self.addTestUserFromLandingPage()
+    landing_page.addTestUser(BaseTest.TEST_USER_AS_DICT['name'],
+                             BaseTest.TEST_USER_AS_DICT['email'],
+                             self.args.server_url)
 
     self.assertTestUserPresenceOnPage(True)
 
   def testDeleteUser(self):
     """Test that deleting a user removes that user."""
-    self.addTestUserFromLandingPage()
+    landing_page = LandingPage(self.driver)
+    landing_page.addTestUser(BaseTest.TEST_USER_AS_DICT['name'],
+                             BaseTest.TEST_USER_AS_DICT['email'],
+                             self.args.server_url)
 
     self.assertTestUserPresenceOnPage(True)
 
-    self.removeTestUser()
+    landing_page.removeTestUser(BaseTest.TEST_USER_AS_DICT['name'],
+                                self.args.server_url)
 
     self.assertTestUserPresenceOnPage(False)
 
   def testDisableAndEnableUserFromListbox(self):
     """Test that disabling and enabling a user works from the listbox."""
-    self.addTestUserFromLandingPage()
+    landing_page = LandingPage(self.driver)
+    landing_page.addTestUser(BaseTest.TEST_USER_AS_DICT['name'],
+                             BaseTest.TEST_USER_AS_DICT['email'],
+                             self.args.server_url)
     self.assertTestUserPresenceOnPage(True)
 
     self.verifyUserCanBeDisabledAndThenEnabled(True, True)
 
   def testDisableAndEnableUserFromDetailsDialog(self):
     """Test that disabling and enabling a user works from the details page."""
-    self.addTestUserFromLandingPage()
+    landing_page = LandingPage(self.driver)
+    landing_page.addTestUser(BaseTest.TEST_USER_AS_DICT['name'],
+                             BaseTest.TEST_USER_AS_DICT['email'],
+                             self.args.server_url)
     self.assertTestUserPresenceOnPage(True)
 
     self.verifyUserCanBeDisabledAndThenEnabled(False, True)
 
   def testCreateNewInviteCode(self):
     """Test that creating a new invite code actually generates a new one."""
+    landing_page = LandingPage(self.driver)
     # Create test user and get it.
-    self.addTestUserFromLandingPage()
+    landing_page.addTestUser(BaseTest.TEST_USER_AS_DICT['name'],
+                             BaseTest.TEST_USER_AS_DICT['email'],
+                             self.args.server_url)
     self.assertTestUserPresenceOnPage(True)
-    test_user_item = self.findTestUserOnLangingPage()
+    test_user_item = landing_page.findTestUser(
+        BaseTest.TEST_USER_AS_DICT['name'])
 
     # Navigate to details dialog.
-    details_dialog = self.getDetailsDialogForItem(test_user_item)
+    details_dialog = landing_page.getDetailsDialogForItem(test_user_item)
 
     # Get the initial invite code.
     initial_invite_code = details_dialog.find_element(
@@ -103,18 +123,19 @@ class LandingPageTest(BaseTest):
     get_invite_code_button.click()
 
     # Wait for post to finish, can take a while.
-    WebDriverWait(self.driver, self.DEFAULT_TIMEOUT).until(
+    WebDriverWait(self.driver, LandingPage.DEFAULT_TIMEOUT).until(
         EC.invisibility_of_element_located(((
             LandingPage.USER_DETAILS_SPINNER))))
 
     # Renavigate to the landing page to ensure the backend changed vs just a
     # UI update if specified.
     self.driver.get(self.args.server_url + flask.url_for('landing'))
-    test_user_item = self.findTestUserOnLangingPage()
+    test_user_item = landing_page.findTestUser(
+        BaseTest.TEST_USER_AS_DICT['name'])
 
     # Set the container element to find the disable/enable button and text on.
     # Will be either the item within the listbox or the overall details dialog.
-    details_dialog = self.getDetailsDialogForItem(test_user_item)
+    details_dialog = landing_page.getDetailsDialogForItem(test_user_item)
 
     # Check the invite code text changed.
     final_invite_code = details_dialog.find_element(
@@ -123,19 +144,31 @@ class LandingPageTest(BaseTest):
 
   def testAddServerFromLandingPage(self):
     """Test that adding a server shows up on the server listing."""
+    landing_page = LandingPage(self.driver)
     self.assertTestServerPresenceOnPage(False)
 
-    self.addTestServerFromLandingPage()
+    landing_page.addTestServer(BaseTest.TEST_SERVER_AS_DICT['ip'],
+                               BaseTest.TEST_SERVER_AS_DICT['name'],
+                               BaseTest.TEST_SERVER_AS_DICT['private_key'],
+                               BaseTest.TEST_SERVER_AS_DICT['public_key'],
+                               self.args.server_url)
 
     self.assertTestServerPresenceOnPage(True)
 
   def testDeleteServer(self):
     """Test that deleting a server removes that server."""
-    self.addTestServerFromLandingPage()
+    landing_page = LandingPage(self.driver)
+    landing_page.addTestServer(BaseTest.TEST_SERVER_AS_DICT['ip'],
+                               BaseTest.TEST_SERVER_AS_DICT['name'],
+                               BaseTest.TEST_SERVER_AS_DICT['private_key'],
+                               BaseTest.TEST_SERVER_AS_DICT['public_key'],
+                               self.args.server_url)
 
     self.assertTestServerPresenceOnPage(True)
 
-    self.removeTestServer()
+    landing_page.removeTestServer(BaseTest.TEST_SERVER_AS_DICT['name'],
+                                  self.args.server_url,
+                                  should_raise_exception=False)
 
     self.assertTestServerPresenceOnPage(False)
 
@@ -155,12 +188,14 @@ class LandingPageTest(BaseTest):
       should_refresh_page: If true, it will refresh the login page in between
                            to check that the backend updates with the UI.
     """
-    test_user_item = self.findTestUserOnLangingPage()
+    landing_page = LandingPage(self.driver)
+    test_user_item = landing_page.findTestUser(
+        BaseTest.TEST_USER_AS_DICT['name'])
 
     # Set the container element to find the disable/enable button and text on.
     # Will be either the item within the listbox or the overall details dialog.
-    container_element = self.getContainerElementForItem(test_user_item,
-                                                        should_use_listbox)
+    container_element = landing_page.getContainerElementForItem(
+        test_user_item, should_use_listbox)
 
     # Get the initial enable/disable text.
     initial_disabled_enabled_text = container_element.find_element(
@@ -169,18 +204,19 @@ class LandingPageTest(BaseTest):
         *LandingPage.USER_DISABLE_ENABLE_BUTTON).text
 
     # Click disable on that user.
-    self.clickDisableEnableOnUserElement(container_element)
+    landing_page.clickDisableEnableOnUserElement(container_element)
 
     # Renavigate to the landing page to ensure the backend changed vs just a
     # UI update if specified.
     if should_refresh_page:
       self.driver.get(self.args.server_url + flask.url_for('landing'))
-      test_user_item = self.findTestUserOnLangingPage()
+      test_user_item = landing_page.findTestUser(
+        BaseTest.TEST_USER_AS_DICT['name'])
 
     # Set the container element to find the disable/enable button and text on.
     # Will be either the item within the listbox or the overall details dialog.
-    container_element = self.getContainerElementForItem(test_user_item,
-                                                        should_use_listbox)
+    container_element = landing_page.getContainerElementForItem(
+        test_user_item, should_use_listbox)
 
     # Check the enable/disable text changed.
     changed_disabled_enabled_text = container_element.find_element(
@@ -193,17 +229,18 @@ class LandingPageTest(BaseTest):
                          changed_disabled_enabled_button_text)
 
     # Flip back to enable.
-    self.clickDisableEnableOnUserElement(container_element)
+    landing_page.clickDisableEnableOnUserElement(container_element)
 
     # Renavigate to the landing page to ensure the backend changed vs just a
     # UI update if specified.
     if should_refresh_page:
       self.driver.get(self.args.server_url + flask.url_for('landing'))
-      test_user_item = self.findTestUserOnLangingPage()
+      test_user_item = landing_page.findTestUser(
+        BaseTest.TEST_USER_AS_DICT['name'])
 
     # Set the container element to find the disable/enable button and text on.
-    container_element = self.getContainerElementForItem(test_user_item,
-                                                        should_use_listbox)
+    container_element = landing_page.getContainerElementForItem(
+        test_user_item, should_use_listbox)
 
     # Check the enable/disable text changed back to the initial.
     final_disabled_enabled_text = container_element.find_element(
@@ -218,59 +255,6 @@ class LandingPageTest(BaseTest):
                       final_disabled_enabled_text)
     self.assertEquals(initial_disabled_enabled_button_text,
                       final_disabled_enabled_button_text)
-
-  def getContainerElementForItem(self, item, should_use_listbox):
-    """Given an item, get the container element for it based on the boolean.
-
-    I realize this method is very basic, but it is just to avoid a lot of
-    duplication across some tests.
-
-    Args:
-      item: The item whose container needs to be returned.
-      should_use_listbox: If true, this will just return the item itself.
-                          Otherwise, this will open and return the details
-                          dialog.
-
-    Returns:
-      The container for the given element based on should_use_listbox, either
-      the item itself or its details dialog once opened.
-    """
-    if should_use_listbox:
-      # Hover the item so the button shows up.
-      ActionChains(self.driver).move_to_element(item).perform()
-      return item
-    else:
-      return self.getDetailsDialogForItem(item)
-
-  def getDetailsDialogForItem(self, item):
-    """Given an item, open it's details dialog and return that element.
-
-    Args:
-      item: The item whose details dialog needs to be opened.
-
-    Returns:
-      The details dialog for the given element, once opened.
-    """
-    item.click()
-    details_modal = item.find_element(*LandingPage.DETAILS_MODAL)
-    WebDriverWait(self.driver, self.DEFAULT_TIMEOUT).until(
-        EC.visibility_of(details_modal))
-    return details_modal
-
-  def clickDisableEnableOnUserElement(self, element):
-    """Given a user element, click the disable/enable button on it.
-
-    Args:
-      element: The user element to click disable/enable upon.
-    """
-    disable_enable_button = element.find_element(
-        *LandingPage.USER_DISABLE_ENABLE_BUTTON)
-    disable_enable_button.click()
-
-    # Wait for post to finish, can take a while.
-    WebDriverWait(self.driver, self.DEFAULT_TIMEOUT).until(
-        EC.invisibility_of_element_located(((
-            LandingPage.USER_DETAILS_SPINNER))))
 
 
 if __name__ == '__main__':
