+++ tests/ui/login_page.py
@@ -6,7 +6,6 @@ from selenium.webdriver.common.by import By
 from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.support.ui import WebDriverWait
 
-from base_test import BaseTest
 from layout import UfOPageLayout
 
 
@@ -52,7 +51,8 @@ class LoginPage(UfOPageLayout):
     """
     self.driver.get(server_url + flask.url_for('login'))
 
-    login_form = WebDriverWait(self.driver, BaseTest.DEFAULT_TIMEOUT).until(
+    login_form = WebDriverWait(self.driver,
+                               UfOPageLayout.DEFAULT_TIMEOUT).until(
         EC.visibility_of_element_located(((self.LOGIN_FORM))))
     username_paper_input = login_form.find_element(*self.USERNAME_INPUT)
     username_input = username_paper_input.find_element(By.ID, 'input')
@@ -66,7 +66,7 @@ class LoginPage(UfOPageLayout):
     sign_in_button.click()
 
     # Wait for redirect to landing page.
-    WebDriverWait(self.driver, BaseTest.DEFAULT_TIMEOUT).until(
+    WebDriverWait(self.driver, UfOPageLayout.DEFAULT_TIMEOUT).until(
         EC.visibility_of_element_located(((LoginPage.OPEN_MENU_BUTTON))))
 
   def Logout(self, server_url):
@@ -76,16 +76,18 @@ class LoginPage(UfOPageLayout):
       server_url: The base url for the server, such as http://0.0.0.0:5000.
     """
     self.driver.get(server_url + flask.url_for('landing'))
-    WebDriverWait(self.driver, BaseTest.DEFAULT_TIMEOUT).until(
+    WebDriverWait(self.driver, UfOPageLayout.DEFAULT_TIMEOUT).until(
         EC.visibility_of_element_located(((LoginPage.OPEN_MENU_BUTTON))))
     dropdown_button = self.driver.find_element(*LoginPage.OPEN_MENU_BUTTON)
     dropdown_button.click()
 
-    logout_form = WebDriverWait(self.driver, BaseTest.DEFAULT_TIMEOUT).until(
+    logout_form = WebDriverWait(self.driver,
+                                UfOPageLayout.DEFAULT_TIMEOUT).until(
         EC.visibility_of_element_located(((self.LOGOUT_FORM))))
     logout_button = logout_form.find_element(*self.GENERIC_PAPER_BUTTON)
     logout_button.click()
 
     # Wait for redirect back to login
-    login_form = WebDriverWait(self.driver, BaseTest.DEFAULT_TIMEOUT).until(
+    login_form = WebDriverWait(self.driver,
+                               UfOPageLayout.DEFAULT_TIMEOUT).until(
         EC.visibility_of_element_located(((self.LOGIN_FORM))))
