+++ tests/ui/setup_page_test.py
@@ -4,6 +4,7 @@ import unittest
 import flask
 
 from base_test import BaseTest
+from landing_page import LandingPage
 from login_page import LoginPage
 from setup_page import SetupPage
 
@@ -21,8 +22,13 @@ class SetupPageTest(BaseTest):
 
   def tearDown(self):
     """Teardown for test methods."""
-    self.removeTestUser(should_raise_exception=False)
-    self.removeTestServer(should_raise_exception=False)
+    landing_page = LandingPage(self.driver)
+    landing_page.removeTestUser(BaseTest.TEST_USER_AS_DICT['name'],
+                                self.args.server_url,
+                                should_raise_exception=False)
+    landing_page.removeTestServer(BaseTest.TEST_SERVER_AS_DICT['name'],
+                                  self.args.server_url,
+                                  should_raise_exception=False)
     LoginPage(self.driver).Logout(self.args.server_url)
     super(SetupPageTest, self).tearDown()
 
@@ -49,17 +55,25 @@ class SetupPageTest(BaseTest):
 
   def testManuallyAddUserFromSetupPage(self):
     """Test that manually adding a user shows up on the user listing."""
+    setup_page = SetupPage(self.driver)
     self.assertTestUserPresenceOnPage(False)
 
-    self.addTestUserFromSetupPage()
+    setup_page.addTestUser(BaseTest.TEST_USER_AS_DICT['name'],
+                           BaseTest.TEST_USER_AS_DICT['email'],
+                           self.args.server_url)
 
     self.assertTestUserPresenceOnPage(True)
 
   def testAddServerFromSetupPage(self):
     """Test that adding a server shows up on the server listing."""
+    setup_page = SetupPage(self.driver)
     self.assertTestServerPresenceOnPage(False)
 
-    self.addTestServerFromSetupPage()
+    setup_page.addTestServer(BaseTest.TEST_SERVER_AS_DICT['ip'],
+                             BaseTest.TEST_SERVER_AS_DICT['name'],
+                             BaseTest.TEST_SERVER_AS_DICT['private_key'],
+                             BaseTest.TEST_SERVER_AS_DICT['public_key'],
+                             self.args.server_url)
 
     self.assertTestServerPresenceOnPage(True)
